+++ javascript/login.js
@@ -46,8 +46,8 @@ function forceLogin() {
     })
       .then(async (res) => {
         const json = await res.json();
-        let txt = ''
-        if (res.status === 200) txt = 'login verified'
+        let txt = '';
+        if (res.status === 200) txt = 'login verified';
         else txt = `${res.status}: ${res.statusText} - ${json.detail}`;
         status.textContent = txt;
         console.log('login', txt);